extract proxyRestrictedTable function
