+++ src/common/SPELLS/index.ts
@@ -12,7 +12,7 @@
  * use the named spell by default, this makes things much more readable.
  **************************************************************************************************************** */
 
-import indexById from 'common/indexById';
+import indexById, { proxyRestrictedTable } from 'common/indexById';
 
 import DEATH_KNIGHT from './deathknight';
 import DEMON_HUNTER from './demonhunter';
@@ -55,42 +55,8 @@ const ABILITIES = {
   ...SHADOWLANDS,
 } as const;
 
-// type SpellCollection = SpellList & {
-//   maybeGet: (key: string | number | undefined) => Spell | undefined;
-// };
-// If you remove this indexById you can see what spells are undefined.
-// But you'll get a lot of other errors.
-// We should type indexById properly some day to make this standard.
-// And then fix all those errors.
-// Which will prevent bugs.
 const InternalSpellTable = indexById<Spell | Enchant, typeof ABILITIES>(ABILITIES);
-// assignment is used here to avoid potential performance pitfalls when
-// compiling the spread operator on large objects.
-// InternalSpellTable.maybeGet = (key) => (key ? InternalSpellTable[key] : undefined);
-
-const SPELLS = new Proxy(InternalSpellTable, {
-  get(target, prop, receiver) {
-    const value = Reflect.get(target, prop, receiver);
-
-    if (value === undefined) {
-      if (process.env.NODE_ENV === 'production') {
-        console.error(
-          'Attempted to retrieve invalid or missing spell from SPELLS. If this is expected, use SPELLS.maybeGet.',
-          prop,
-          target,
-        );
-      } else {
-        throw new Error(
-          `Attempted to retrieve invalid or missing spell from SPELLS: ${String(
-            prop,
-          )}. If this is expected, use SPELLS.maybeGet.`,
-        );
-      }
-    }
-
-    return value;
-  },
-});
+const SPELLS = proxyRestrictedTable(InternalSpellTable, 'SPELLS', 'maybeGetSpell');
 
 export default SPELLS;
