+++ src/common/indexById.ts
@@ -11,10 +11,15 @@ type RestrictedTable<T, E> = {
   [Key in keyof E]: E[Key] extends T ? E[Key] : never;
 };
 
+type IndexedRestrictedTable<
+  ValueT extends BaseIndexableObj,
+  Map extends RestrictedTable<ValueT, any>
+> = RestrictedTable<ValueT, Map> & Record<number, ValueT>;
+
 const indexById = <ValueT extends BaseIndexableObj, Map extends RestrictedTable<ValueT, any>>(
   arg: Map,
-): RestrictedTable<ValueT, Map> & Record<number, ValueT> => {
-  const indexedByNameAndId: RestrictedTable<ValueT, Map> & Record<number, ValueT> = { ...arg };
+): IndexedRestrictedTable<ValueT, Map> => {
+  const indexedByNameAndId: IndexedRestrictedTable<ValueT, Map> = { ...arg };
   typedKeys(arg).forEach((key) => {
     const value = arg[key];
 
@@ -65,6 +70,39 @@ export const indexOnlyById = <
     indexedByNameAndId[value.id] = value;
   });
   return indexedByNameAndId;
+};
+
+export const proxyRestrictedTable = <
+  ValueT extends BaseIndexableObj,
+  Map extends RestrictedTable<ValueT, any>
+>(
+  restrictedTable: IndexedRestrictedTable<ValueT, Map>,
+  valueBeingAccessed: string,
+  alternative: string,
+): IndexedRestrictedTable<ValueT, Map> => {
+  return new Proxy<IndexedRestrictedTable<ValueT, Map>>(restrictedTable, {
+    get(target, prop, receiver) {
+      const value = Reflect.get(target, prop, receiver);
+
+      if (value === undefined) {
+        if (process.env.NODE_ENV === 'production') {
+          console.error(
+            `Attempted to retrieve invalid or missing spell from ${valueBeingAccessed}. If this is expected, use ${alternative}.`,
+            prop,
+            target,
+          );
+        } else {
+          throw new Error(
+            `Attempted to retrieve invalid or missing spell from ${valueBeingAccessed}: ${String(
+              prop,
+            )}. If this is expected, use ${alternative}.`,
+          );
+        }
+      }
+
+      return value;
+    },
+  });
 };
 
 export default indexById;
